Generating 1000 scenarios and return the one with the highest closrPrice at the end of the last month implemented.
